third_party/WebKit/LayoutTests/http/tests/security/xssAuditor/xss-filter-bypass-big5-expected.txt

CONSOLE ERROR: line 4: The XSS Auditor refused to execute a script in 'http://localhost:8000/security/xssAuditor/resources/echo-intertag.pl?q=%3Cscript%3Ealert(%27%b4%5f%27)%3C/script%3E&charset=big5&notifyDone=1' because its source code was found within the request. The server sent an 'X-XSS-Protection' header requesting this behavior.
